moved utility functions to standalone class
