+++ archive_upload/handlers/dsmc_handlers.py
@@ -15,6 +15,7 @@ from arteria.web.handlers import BaseRestHandler
 
 from archive_upload import __version__ as version
 from archive_upload.lib.jobrunner import LocalQAdapter
+from archive_upload.lib.utils import FileUtils
 
 from tornado import web
 
@@ -87,26 +88,6 @@ class BaseDsmcHandler(BaseRestHandler):
         if len(files) == 0 and remove_root:
             log.info("Removing empty folder: {}".format(path))
             os.rmdir(path)
-
-    @staticmethod
-    def _list_all_paths(path, followlinks=False):
-        """
-        Recursively traverse the supplied folder with os.walk and return a list of full paths to all files and folders
-        beneath the path. Essentially equivalent to `find path`. Note that symlinks pointing to files are
-        returned as well.
-
-        The returned list will be sorted in reverse lexical order, meaning that paths in subdirectories will come
-        before the parent directories
-
-        :param path: folder, beneath which to list all files
-        :param followlinks: if True, follow symlinks to directories (default False)
-        :return: a list of full paths discovered with os.walk
-        """
-
-        all_paths = []
-        for dirpath, subdirs, dirfiles in os.walk(os.path.normpath(path), followlinks=followlinks):
-            all_paths.extend(map(lambda f: os.path.join(dirpath, f), dirfiles + subdirs))
-        return sorted(all_paths, reverse=True)
 
     def write_error(self, status_code, **kwargs):
         self.set_header("Content-Type", "application/json")
@@ -699,40 +680,6 @@ class CompressArchiveHandler(BaseDsmcHandler):
     Handler for compressing certain files in the archive before uploading.
     """
 
-    @staticmethod
-    def source_paths_from_tarball(tarball, path_to_source):
-        """
-        List the paths inside the tarball and return their full paths rooted at the supplied source path
-
-        :param tarball: the tarball to list files from
-        :param path_to_source: the path to the root of the source folder
-        :return: a list of the paths inside the tarball, using the supplied source path as root
-        """
-        with tarfile.open(tarball) as tar:
-            return map(
-                lambda m: os.path.normpath(os.path.join(path_to_source, m.name)),
-                tar.getmembers())
-
-    @staticmethod
-    def paths_duplicated_in_tarball(tarball, path_to_archive):
-        """
-        List the files and folders in the supplied folder and its subdirectories and return the paths
-        that are present in the supplied tarball, assuming that the tarball is rooted
-        at the supplied folder.
-
-        :param tarball: a tarball whose members are rooted at the supplied path
-        :param path_to_archive: path to search for files and folders duplicated in the tarball
-        :return: a list of duplicated files and folders, sorted in reverse lexical order
-        """
-        # list the paths in the tarball and store as a set
-        paths_in_tarball = set(CompressArchiveHandler.source_paths_from_tarball(tarball, path_to_archive))
-        paths_in_source_archive = BaseDsmcHandler._list_all_paths(path_to_archive, followlinks=False)
-
-        # duplicated paths are present in tarball and on disk, so take the intersection of the lists
-        duplicated_paths = paths_in_tarball.intersection(paths_in_source_archive)
-
-        return sorted(list(duplicated_paths), reverse=True)
-
     def post(self, archive):
         """
         Create a gziped tarball of most files in the archive, with the exception of
@@ -791,7 +738,7 @@ class CompressArchiveHandler(BaseDsmcHandler):
         # Remove files that were added to the tarball
 
         # get a list of files and folders duplicated between the tarball and the archive
-        paths_to_remove = CompressArchiveHandler.paths_duplicated_in_tarball(tarball_path, path_to_archive)
+        paths_to_remove = FileUtils.paths_duplicated_in_tarball(tarball_path, path_to_archive)
 
         # remove the files first
         for file_to_remove in filter(os.path.isfile, paths_to_remove):
